requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.24.1
google-generativeai==0.3.0
faiss-cpu==1.8.0
packaging>=21.0
pdfplumber==0.10.3
python-dotenv==1.0.0
aiofiles==23.2.1
PyPDF2==3.0.1
tiktoken==0.5.2
numpy==1.26.4
scikit-learn==1.3.2
